Generated by Spark: I need a simple app where people answer a "This or That" questions, and the items being compared are predetermined. And their comparison happens like a disqualification way.
For example, consider three items I mention:
1. iPhone
2. Airwrap
3. Dark Chocolate

The app should first present the user with
Dark Chocolate ot Airwrap

If the user selected Airwrap, then it should eliminate Dark Chocolate and ask "iPhone or Airwrap"?

There can be any number of items here, but we need to use elimination criteria to reach the last one.

Idea is to present users with a quiz, and understand something they want.